# server/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/auth/register": {
      "post": {
        "operationId": "AuthController_register",
        "summary": "Inscription d'un nouvel utilisateur",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Utilisateur inscrit",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Erreur de validation ou utilisateur existant"
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "summary": "Connexion utilisateur",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Connexion réussie",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Identifiants invalides"
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/refresh": {
      "post": {
        "operationId": "AuthController_refresh",
        "summary": "Rafraîchissement du token d'accès",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshTokenDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Nouveaux tokens générés",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Token de rafraîchissement invalide ou expiré"
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/onboarding": {
      "post": {
        "operationId": "AuthController_completeOnboarding",
        "summary": "Compléter le questionnaire d'onboarding",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OnboardingDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Profil utilisateur mis à jour avec succès."
          },
          "400": {
            "description": "Données invalides."
          },
          "404": {
            "description": "Utilisateur non trouvé."
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/profile": {
      "get": {
        "operationId": "AuthController_getProfile",
        "summary": "Récupérer le profil de l'utilisateur connecté",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Profil utilisateur retourné",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Auth"
        ]
      },
      "put": {
        "operationId": "AuthController_updateProfile",
        "summary": "Mettre à jour le profil de l'utilisateur connecté",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProfileDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Profil mis à jour avec succès",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Données invalides ou erreur de validation"
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/change-password": {
      "post": {
        "operationId": "AuthController_changePassword",
        "summary": "Changer le mot de passe de l'utilisateur connecté",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Mot de passe mis à jour avec succès"
          },
          "400": {
            "description": "Données invalides ou erreur de validation"
          },
          "401": {
            "description": "Mot de passe actuel incorrect"
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/settings": {
      "get": {
        "operationId": "AuthController_getSettings",
        "summary": "Récupérer tous les paramètres de l'utilisateur connecté",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Paramètres utilisateur récupérés avec succès",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SettingsResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Auth"
        ]
      },
      "put": {
        "operationId": "AuthController_updateSettings",
        "summary": "Mettre à jour les paramètres de l'utilisateur connecté",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSettingsDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Paramètres mis à jour avec succès",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateSettingsResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Données invalides ou erreur de validation"
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
...
    "/auth/preferences/workouts": {
      "put": {
        "operationId": "AuthController_updateWorkoutPreferences",
        "summary": "Mettre à jour les préférences d'entraînement",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateWorkoutPreferencesDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Préférences d'entraînement mises à jour avec succès",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateSettingsResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Données invalides ou erreur de validation"
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/preferences/notifications": {
      "put": {
        "operationId": "AuthController_updateNotificationPreferences",
        "summary": "Mettre à jour les préférences de notifications",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateNotificationPreferencesDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Préférences de notifications mises à jour avec succès",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateSettingsResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Données invalides ou erreur de validation"
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
...
  },
  "info": {
    "title": "FitMe API",
    "description": "Documentation de l'API FitMe - Application de fitness basée sur le cycle menstruel",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "RegisterDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "user@example.com",
            "description": "Adresse email"
          },
          "password": {
            "type": "string",
            "example": "motdepasse123",
            "minLength": 8,
            "description": "Mot de passe (min 8 caractères)"
          },
          "firstName": {
            "type": "string",
            "example": "Jane",
            "description": "Prénom"
          },
          "profileType": {
            "type": "string",
            "example": "FEMALE",
            "description": "Type de profil"
          },
          "contextType": {
            "type": "string",
            "example": "CYCLE",
            "description": "Type de contexte"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "AuthResponseDto": {
        "type": "object",
        "properties": {
          "tokens": {
            "type": "object",
            "properties": {
              "accessToken": {
                "type": "string",
                "example": "eyJhbGciOiJIUzI1NiIs..."
              },
              "refreshToken": {
                "type": "string",
                "example": "eyJhbGciOiJIUzI1NiIs..."
              },
              "expiresIn": {
                "type": "number",
                "example": 900
              }
            }
          },
          "user": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string",
                "example": "cuid_example"
              },
              "email": {
                "type": "string",
                "example": "user@example.com"
              },
              "firstName": {
                "type": "string",
                "example": "Jane"
              },
              "profileType": {
                "type": "string",
                "example": "FEMALE"
              },
              "contextType": {
                "type": "string",
                "example": "CYCLE"
              },
              "onboardingCompleted": {
                "type": "boolean",
                "example": false
              },
              "experienceLevel": {
                "type": "string",
                "example": "INTERMEDIATE"
              }
            }
          }
        },
        "required": [
          "tokens",
          "user"
        ]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "user@example.com",
            "description": "Adresse email"
          },
          "password": {
            "type": "string",
            "example": "motdepasse123",
            "description": "Mot de passe"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "RefreshTokenDto": {
        "type": "object",
        "properties": {
          "refreshToken": {
            "type": "string",
            "example": "eyJhbGciOiJIUzI1NiIs...",
            "description": "Token de rafraîchissement JWT"
          }
        },
        "required": [
          "refreshToken"
        ]
      },
      "TokenResponseDto": {
        "type": "object",
        "properties": {
          "tokens": {
            "type": "object",
            "properties": {
              "accessToken": {
                "type": "string",
                "example": "eyJhbGciOiJIUzI1NiIs..."
              },
              "refreshToken": {
                "type": "string",
                "example": "eyJhbGciOiJIUzI1NiIs..."
              },
              "expiresIn": {
                "type": "number",
                "example": 900
              }
            }
          }
        },
        "required": [
          "tokens"
        ]
      },
      "OnboardingDto": {
        "type": "object",
        "properties": {
          "objective": {
            "type": "string",
            "enum": [
              "WEIGHT_LOSS",
              "MUSCLE_GAIN",
              "ENDURANCE",
              "STRENGTH",
              "FLEXIBILITY",
              "GENERAL_FITNESS",
              "STRESS_REDUCTION",
              "ENERGY_BOOST"
            ],
            "example": "GENERAL_FITNESS",
            "description": "Objectif principal de l'utilisateur"
          },
          "experienceLevel": {
            "type": "string",
            "enum": [
              "BEGINNER",
              "INTERMEDIATE",
              "ADVANCED"
            ],
            "example": "INTERMEDIATE",
            "description": "Niveau d'expérience sportive de l'utilisateur"
          },
          "isMenopausal": {
            "type": "boolean",
            "example": false,
            "description": "Indique si l'utilisatrice est en ménopause"
          },
          "averageCycleLength": {
            "type": "number",
            "example": 28,
            "description": "Durée moyenne du cycle menstruel en jours"
          },
          "averagePeriodLength": {
            "type": "number",
            "example": 5,
            "description": "Durée moyenne des règles en jours"
          }
        },
        "required": [
          "objective",
          "experienceLevel",
          "isMenopausal"
        ]
      },
      "ProfileResponseDto": {
        "type": "object",
        "properties": {
          "user": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string",
                "example": "cuid_example"
              },
              "email": {
                "type": "string",
                "example": "user@example.com"
              },
              "firstName": {
                "type": "string",
                "example": "Jane"
              },
              "birthDate": {
                "type": "string",
                "example": "1990-05-15"
              },
              "profileType": {
                "type": "string",
                "example": "FEMALE"
              },
              "contextType": {
                "type": "string",
                "example": "CYCLE"
              },
              "objective": {
                "type": "string",
                "example": "WEIGHT_LOSS"
              },
              "sportFrequency": {
                "type": "string",
                "example": "3_TIMES_WEEK"
              },
              "isMenopausal": {
                "type": "boolean",
                "example": false
              }
            }
          }
        },
        "required": [
          "user"
        ]
      },
      "UpdateProfileDto": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string",
            "example": "Jane",
            "description": "Prénom"
          },
          "birthDate": {
            "type": "string",
            "example": "1990-05-15",
            "description": "Date de naissance (format YYYY-MM-DD)"
          },
          "profileType": {
            "type": "string",
            "example": "FEMALE",
            "description": "Type de profil (FEMALE, MALE, NON_BINARY, OTHER)",
            "enum": [
              "FEMALE",
              "MALE",
              "NON_BINARY",
              "OTHER"
            ]
          },
          "contextType": {
            "type": "string",
            "example": "CYCLE",
            "description": "Type de contexte (CYCLE, GENERAL, MENOPAUSE, NONE)",
            "enum": [
              "CYCLE",
              "GENERAL",
              "MENOPAUSE",
              "NONE"
            ]
          },
          "objective": {
            "type": "string",
            "example": "WEIGHT_LOSS",
            "description": "Objectif principal"
          },
          "sportFrequency": {
            "type": "string",
            "example": "3_TIMES_WEEK",
            "description": "Fréquence de sport"
          },
          "isMenopausal": {
            "type": "boolean",
            "example": false,
            "description": "Statut ménopause"
          }
        }
      },
      "ChangePasswordDto": {
        "type": "object",
        "properties": {
          "currentPassword": {
            "type": "string",
            "example": "ancienMotDePasse123",
            "description": "Mot de passe actuel"
          },
          "newPassword": {
            "type": "string",
            "example": "nouveauMotDePasse123",
            "minLength": 8,
            "description": "Nouveau mot de passe (min 8 caractères)"
          }
        },
        "required": [
          "currentPassword",
          "newPassword"
        ]
      },
...
    }
  }
}